Current Expected Credit Loss
To test the FASB’s new framework CECL which is forward-looking, we choose Fannie Mae single-family data for the state of California as it’s the most volatile and 22% of the Fannie Mae data. To evaluate this, we choose the vintage model, but it has a high error rate. As the data is unbalanced, then we choose PD x LGD methodology. For the PD model, we used a transition matrix including 11 macroeconomic variables and for the LGD model, we used Fractional Response regression.